+++ Constraid/ManageSize.swift
@@ -49,6 +49,28 @@ public func expandWidth(of item: Constraid.View, from constant: CGFloat, priorit
 }
 
 /**
+ Limit width of receiver using a constraint in auto-layout
+
+ - parameter item: The `item` you want to constrain
+ - parameter constant: The maximum width to limit by
+ - parameter priority: The priority this constraint uses when being
+ evaluated against other constraints
+
+ - returns: Constraint collection containing the generated constraint
+ */
+@discardableResult
+public func limitWidth(of item: Constraid.View, by constant: CGFloat, priority: Constraid.LayoutPriority = Constraid.LayoutPriorityRequired) -> Constraid.ConstraintCollection {
+
+    item.translatesAutoresizingMaskIntoConstraints = false
+    let collection = Constraid.ConstraintCollection([
+        NSLayoutConstraint(item: item, attribute: .width,
+                           relatedBy: .lessThanOrEqual, toItem: nil, attribute: .notAnAttribute,
+                           multiplier: 1.0, constant: constant, priority: priority)
+        ])
+    return collection
+}
+
+/**
  Set height of receiver using a constraint in auto-layout
 
  - parameter item: The `item` you want to constrain